pyproject.toml
[project]
name = "wetter-berg"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "discord-py>=2.6.4",
    "dotenv>=0.9.9",
    "pydantic>=2.12.5",
    "requests>=2.32.5",
    "requests-cache>=1.2.1",
    "retry-requests>=2.0.0",
]

[dependency-groups]
dev = [
    "types-requests>=2.32.4.20250913",
]
